requirements.txt
click==*
PyYAML==*
